+++ src/pages/Login/index.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from '@/features/Login/Components';
+import { LoginForm } from '@/features/Login/ui';
 
 export default function LoginPage() {
   return <LoginForm />;
