+++ src/pages/SignUp/index.tsx
@@ -1,4 +1,4 @@
-import { SignUpForm } from '@/features/SignUp/Components';
+import { SignUpForm } from '@/features/SignUp/ui';
 
 export default function SignUpPage() {
   return <SignUpForm />;
